Use a new process to host http server
Before this patch, we are using a new thread to
host the http server. It works in most cases, but hit
the python3.5 partial writes bug (if the script is
run by python3.5), which lead to firmware image
downloading failure on BMC side.
The python3.5 partial writes bug:
  Issue: https://github.com/python/cpython/issues/70908
  Fix: https://github.com/python/cpython/commit/34eeed42901666fce099947f93dfdfc05411f286#diff-c00d56a0c132ee4bdf79f55a4b43643cf314df1fe122c07c539e120c7ec98b5e

That python3.5 bug can be reproduced in a very rigorous env.
In our script case, there are two kinds of operations:
   1: Script have a thread to host http server, BMC will http download
   Firmware image from that http server.
   2: At the same time, Script will also send redfish(http) request
   to BMC, to poll the downloading status.

   These two operations(traffic) need:
   a: happen at the same time.
   b: happen within the same process (different thread).

   For #a: That's why downloading image manually by wget or curl
           do not reproduce the downloading failure issue
   For #b: That's how we fix the problem. We host the http server
           in a new process, to avoid hitting that python3.5 bug.