Wait for DPU ready for upto 3 minutes
Before this patch, it poll the DPU state for upto 100 times.
for each time, it may take about ~2(redfish) + 4(sleep) seconds. That means
it will take about 10 minutes to timeout. The timeout value depends
on the timeout used on redfish command.

After this patch, the timeout value is exactly 3 minutes. It does not
depend on the redfish command.